requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
idna==3.4
Pillow==9.5.0
python-magic-bin==0.4.14
requests==2.30.0
urllib3==2.0.2